requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
pytesseract==0.3.10
Pillow==10.1.0
rapidfuzz==3.5.2
sentence-transformers>=2.3.0
faiss-cpu==1.7.4
openai>=1.12.0
pandas==2.1.3
numpy==1.26.2
pytest==7.4.3
python-dotenv==1.0.0


